requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
dnspython==2.3.0
environs==9.5.0
filelock==3.12.0
grpcio==1.53.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
motor==3.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pika==1.3.2
Pillow==9.5.0
protobuf==4.23.0
pymilvus==2.2.8
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sympy==1.11.1
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.2
